backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
numpy==1.26.3
pandas==2.2.0
yfinance==0.35.0
tushare==1.4.14
redis==5.0.1
websockets==12.0
python-dotenv==1.0.0
pytest==7.4.4
httpx==0.26.0
